Migrate to go-gh text package
