+++ pkg/cmd/repo/deploy-key/list/list.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/cli/cli/v2/internal/ghrepo"
+	"github.com/cli/cli/v2/internal/text"
 	"github.com/cli/cli/v2/pkg/cmdutil"
 	"github.com/cli/cli/v2/pkg/iostreams"
 	"github.com/cli/cli/v2/utils"
@@ -81,7 +82,7 @@ func listRun(opts *ListOptions) error {
 
 		createdAt := deployKey.CreatedAt.Format(time.RFC3339)
 		if t.IsTTY() {
-			createdAt = utils.FuzzyAgoAbbr(now, deployKey.CreatedAt)
+			createdAt = text.FuzzyAgoAbbr(now, deployKey.CreatedAt)
 		}
 		t.AddField(createdAt, nil, cs.Gray)
 		t.EndRow()
