+++ pkg/cmd/ssh-key/list/list.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"github.com/cli/cli/v2/internal/config"
+	"github.com/cli/cli/v2/internal/text"
 	"github.com/cli/cli/v2/pkg/cmdutil"
 	"github.com/cli/cli/v2/pkg/iostreams"
 	"github.com/cli/cli/v2/utils"
@@ -72,7 +73,7 @@ func listRun(opts *ListOptions) error {
 
 		createdAt := sshKey.CreatedAt.Format(time.RFC3339)
 		if t.IsTTY() {
-			createdAt = utils.FuzzyAgoAbbr(now, sshKey.CreatedAt)
+			createdAt = text.FuzzyAgoAbbr(now, sshKey.CreatedAt)
 		}
 		t.AddField(createdAt, nil, cs.Gray)
 		t.EndRow()
