--- pkg/text/indent.go
@@ -1,15 +1,0 @@
-package text
-
-import (
-	"regexp"
-	"strings"
-)
-
-var lineRE = regexp.MustCompile(`(?m)^`)
-
-func Indent(s, indent string) string {
-	if len(strings.TrimSpace(s)) == 0 {
-		return s
-	}
-	return lineRE.ReplaceAllLiteralString(s, indent)
-}
