+++ utils/table_printer.go
@@ -6,8 +6,8 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/cli/cli/v2/internal/text"
 	"github.com/cli/cli/v2/pkg/iostreams"
-	"github.com/cli/cli/v2/pkg/text"
 )
 
 type TablePrinter interface {
