--- utils/terminal.go
@@ -1,25 +1,0 @@
-package utils
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/mattn/go-isatty"
-	"golang.org/x/term"
-)
-
-var IsTerminal = func(f *os.File) bool {
-	return isatty.IsTerminal(f.Fd()) || IsCygwinTerminal(f)
-}
-
-func IsCygwinTerminal(f *os.File) bool {
-	return isatty.IsCygwinTerminal(f.Fd())
-}
-
-var TerminalSize = func(w interface{}) (int, int, error) {
-	if f, isFile := w.(*os.File); isFile {
-		return term.GetSize(int(f.Fd()))
-	}
-
-	return 0, 0, fmt.Errorf("%v is not a file", w)
-}
